backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-Limiter==3.5.0
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
gunicorn==21.2.0
